move get_max_l1_size to common functions
