+++ ttnn/cpp/ttnn/operations/data_movement/untilize_with_unpadding/device/untilize_with_unpadding_program_factory.cpp
@@ -14,6 +14,7 @@
 #include <tt-metalium/host_api.hpp>
 #include "ttnn/common/constants.hpp"
 #include "ttnn/operation.hpp"
+#include "ttnn/operations/functions.hpp"
 
 using namespace tt::constants;
 using namespace tt::tt_metal;
@@ -30,20 +31,12 @@ inline uint32_t get_estimated_size_of_cbs(
     return cb_src0_size + cb_output_size;
 }
 
-inline uint32_t get_max_l1_space(const Tensor& input_tensor_a) {
-    auto device = input_tensor_a.device();
-    auto lowest_address = device->lowest_occupied_compute_l1_address();
-    uint32_t max_l1_space = lowest_address.has_value() ? lowest_address.value() : device->l1_size_per_core();
-    max_l1_space = max_l1_space - device->get_base_allocator_addr(HalMemType::L1);
-    return max_l1_space;
-}
-
 inline bool enough_available_space(
     const Tensor& input_tensor_a,
     const uint32_t input_single_tile_size,
     const uint32_t output_single_tile_size,
     const uint32_t num_tiles_per_row) {
-    uint32_t max_l1_space = get_max_l1_space(input_tensor_a);
+    uint32_t max_l1_space = ttnn::get_max_l1_space(input_tensor_a);
     uint32_t estimated_size_of_cbs =
         get_estimated_size_of_cbs(input_tensor_a, input_single_tile_size, output_single_tile_size, num_tiles_per_row);
     return max_l1_space > estimated_size_of_cbs;
